Changed Grant.otherFunders list to single Grant.directFunder
Out of a discussion with @jrmartino. Each funding source is a new "grant" in the system i.e. grants have a one to one relationship with the original funder.  A grant can, however, have an intermediary organization who receives the funding and then "writes the checks" to other institutions.  For example, where co-pis are in different institutions. This change reflects this - all records should have a primaryFunder and a directFunder that indicate where the money came from and who wrote the check, respectively. These will be the same in most cases.